Staging

+++ SmartDigitalPsico.Domain/DTO/Medical/MedicalCalendar/ActionMedicalCalendarDtoBase.cs
@@ -32,7 +32,7 @@ namespace SmartDigitalPsico.Domain.DTO.Medical.MedicalCalendar
         public DayOfWeek[] RecurrenceDays { get; set; } = [];
         public ERecurrenceCalendarType RecurrenceType { get; set; }
         public DateTime? RecurrenceEndDate { get; set; }
-        public byte RecurrenceCount { get; set; }
+        public short RecurrenceCount { get; set; }
         #endregion Columns  
     }
 }
